Introduce a notion of time, so we can log
As the telegrams of the Kamstrup do not contain a proper timestamp
we're forced to log against system time.

The microcontroller has no RTC, but we don't need that precision
anyways.

So, here's what I'm doing now:
- on setup, get the time from an NTP server
- on every telegram processing, take the system time as the timestamp
- update time every 5 minutes from ntp (could probably be less)

As the time library was already in use, this doesn't add much in size.

What we do need to do though is writing the systemtime based on the
timestamp (we let the ntp syncer do that). In the original, the
"system" time followed the timestamps of the telegrams.

I'm not away of any memory in the reader, so there's not much
difference between the two methods.